+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- Fix issues with account linking [#11](https://github.com/rokwire/app-flutter-plugin/issues/11)
 - Allow more functions in services to be overridden [#1](https://github.com/rokwire/app-flutter-plugin/issues/1)
 - Added RoundedButton to UI section [#2](https://github.com/rokwire/app-flutter-plugin/issues/2).
 - Added header bars to UI section [#2](https://github.com/rokwire/app-flutter-plugin/issues/2).